feat: prevent make dummy proposal usage
